+++ plugins/dates/tests/tokenizer.test.js
@@ -10,7 +10,7 @@ test('date-tokenizer', function (t) {
     ['monday, wednesday, and friday', 3],
     ['between monday and friday', 1],
     ['june to august', 1],
-    ['june or august', 2],
+    ['june or august', 1],
     ['june through august', 1],
     ['june 2020 or august', 2],
     ['mon tue fri', 3],