+++ plugins/dates/changelog.md
@@ -1,5 +1,5 @@
 <!-- #### [Unreleased]
-
+- **[breaking]** - return array in .json().dates
 -->
 
 ### 2.1.0 [April 2021]